requirements.txt
grpcio==1.66.2
grpcio-tools==1.66.2
protobuf==5.28.2
random-username==1.0.2
